\name{functions}

\alias{belegundu}
\alias{belegundu.constr}
\alias{binh1}
\alias{binh2}
\alias{binh2.constr}
\alias{binh3}
\alias{deb3}
\alias{fonseca1}
\alias{fonseca2}
\alias{gianna}
\alias{hanne1}
\alias{hanne1.constr}
\alias{hanne2}
\alias{hanne2.constr}
\alias{hanne3}
\alias{hanne3.constr}
\alias{hanne4}
\alias{hanne4.constr}
\alias{hanne5}
\alias{hanne5.constr}
\alias{jimenez}
\alias{jimenez.constr}
\alias{vnt}
\alias{zdt1}
\alias{zdt2}
\alias{zdt3}

\title{MCO test problems}
\description{
  Collection of functions implementing various MCO test problems.
}
\usage{
belegundu(x)
belegundu.constr(x)
binh1(x)
binh2(x)
binh2.constr(x)
binh3(x)
deb3(x)
fonseca1(x)
fonseca2(x)
gianna(x)
hanne1(x)
hanne1.constr(x)
hanne2(x)
hanne2.constr(x)
hanne3(x)
hanne3.constr(x)
hanne4(x)
hanne4.constr(x)
hanne5(x)
hanne5.constr(x)
jimenez(x)
jimenez.constr(x)
vnt(x)
zdt1(x)
zdt2(x)
zdt3(x)
}
\arguments{
  \item{x}{Input vector}
}
\value{
}
\references{ ~put references to the literature/web site here ~ }
\author{
  Heike Trautmann \email{trautmann@statistik.tu-dortmund.de},
  Detlef Steuer \email{steuer@hsu-hamburg.de} and
  Olaf Mersmann \email{olafm@statistik.tu-dortmund.de}
}
\examples{
}
