\name{paretoFront}
\alias{paretoFront}
\alias{paretoSet}
\title{Pareto Front and pareto set accessor}
\description{
  Extract the pareto front or pareto set from an mco result object.
}
\usage{
paretoFront(x, ...)
paretoSet(x, ...)
}
\arguments{
  \item{x}{Result of an mco run}
  \item{\dots}{Ignored}
}
\value{
  The pareto front or pareto set.
}
\author{
  Heike Trautmann \email{trautmann@statistik.uni-dortmund.de},
  Detlef Steuer \email{steuer@hsu-hamburg.de} and
  Olaf Mersmann \email{olafm@statistik.uni-dortmund.de}
}
